fix(defaults): handle list-descriptions of options
