+++ tests/fixtures/roles/generate_defaults/meta/argument_specs.yml
@@ -49,7 +49,9 @@ argument_specs:
           - value1
           - value2
           - value3
-        description: A list value
+        description:
+          - A list value
+          - with a list description with a rather long description, maybe too long.
       generate_defaults_dict:
         type: dict
         default:
